backend/telegram-sender/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
pytz==2023.3
openai==1.12.0